requirements.txt
aioredis==1.3.1
2captcha-python==1.1.3
apns2==0.7.1
arrow==1.2.3
asgiref==3.8.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cachetools==5.4.0
certifi==2025.4.26
cffi==1.14.5
channels==3.0.3
channels-redis==3.4.1
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cssutils==2.2.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.6
dj-database-url==0.5.0
dj-rest-auth==6.0.0
Django==3.2
django-allauth==0.55.2
django-appconf==1.0.5
django-compressor==4.3.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-dotenv==1.4.2
django-encrypted-model-fields==0.6.5
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.7.1
django-inlinecss==0.3.0
django-nested-admin==3.4.0
django-rest-passwordreset==1.3.0
django-rest-auth==0.9.5
django-storages==1.11.1
django-user-agents==0.4.0
djangochannelsrestframework==1.1.0
django-redis==5.2.0
djangorestframework==3.15.1
drf-extra-fields==3.4.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.12.0
fake-useragent==1.4.0
future==0.18.2
gunicorn==20.0.4
h2==2.6.2
hiredis==2.0.0
hpack==3.0.0
http_ece==1.1.0
httplib2==0.22.0
hyper==0.7.0
hyperframe==3.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflect==6.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==4.17.3
MarkupPy==1.14
MarkupSafe==1.1.1
msgpack==1.0.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pipenv==2023.4.20
platformdirs==3.2.0
protobuf==4.25.3
py-vapid==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.20.0
pynliner==0.8.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.20.0
python-dateutil==2.8.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.1
pywebpush==1.14.0
PyYAML==6.0
rcssmin==1.1.1
requests==2.25.1
redis==4.5.4
requests-oauthlib==1.3.0
rjsmin==1.2.1
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
selenium==4.15.2
service-identity==21.1.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
tablib==3.2.1
TatSu==5.8.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.3.0
ua-parser==0.16.1
uritemplate==3.0.1
urllib3==1.26.3
undetected-chromedriver==3.5.4
user-agents==2.2.0
useragent==0.1.1
virtualenv==20.22.0
virtualenv-clone==0.5.7
websockets==10.3
webdriver-manager==4.0.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.1
zope.interface==5.4.0
psycopg2==2.9.10
